requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.6.0
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cssselect==1.1.0
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.3.0/en_core_web_lg-3.3.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.3
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
langcodes==3.3.0
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
murmurhash==1.0.7
mypy-extensions==0.4.3
nest-asyncio==1.5.5
nltk==3.7
numpy==1.23.0
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
preshed==3.0.6
prompt-toolkit==3.0.30
protobuf==4.21.2
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==23.2.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.3.0
thinc==8.0.17
thinc-apple-ops==0.0.7
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.12.0
torchtext==0.13.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typer==0.4.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
